kv/kvserver: clarify declared keys for RequestLease request
The request type does not acquire latches but does still declare keys, which
was confusing without an accompanying comment.
